added the "delete restaurant" function to admin

+++ backofficeModifyUser.php
@@ -11,12 +11,23 @@
     $rows = mysql_num_rows($query);
     if ($rows != 0) {
         while ($row = mysql_fetch_assoc($query)) {
-            ?>
-            <tr>
-                <td><?php echo($row['userName']); ?></td>
-                <td><a href="/tools/backofficeDeleteUser.php?userID=<?php echo($row['userID']); ?>">Delete user</a></td>
-            </tr>
-        <?php
+            if ($row['userID'] != $_SESSION['userID']) {
+                ?>
+                <tr>
+                    <td><?php echo($row['userName']); ?></td>
+                    <td>
+                        <?php
+                        if($row['isAdmin']==1){
+                            echo("Admin");
+                        }else{
+                            echo("user");
+                        }
+                        ?>
+                    </td>
+                    <td><a href="/tools/backofficeDeleteUser.php?userID=<?php echo($row['userID']); ?>">Delete user</a></td>
+                </tr>
+            <?php
+            }
         }
     }?>
 </table>
